Setup Artificial Intelligence editing as sub navigation
