+++ app-modules/consent/src/Filament/Resources/ConsentAgreementResource/Pages/EditConsentAgreement.php
@@ -2,14 +2,8 @@
 
 namespace Assist\Consent\Filament\Resources\ConsentAgreementResource\Pages;
 
-use Filament\Forms\Form;
-use Filament\Actions\Action;
-use Filament\Forms\Components\Fieldset;
-use Filament\Forms\Components\Textarea;
-use Filament\Forms\Components\TextInput;
 use Filament\Resources\Pages\EditRecord;
 use Illuminate\Contracts\Support\Htmlable;
-use Assist\Consent\Models\UserConsentAgreement;
 use Assist\Consent\Filament\Resources\ConsentAgreementResource;
 
 class EditConsentAgreement extends EditRecord
@@ -19,49 +13,5 @@ class EditConsentAgreement extends EditRecord
     public function getTitle(): string | Htmlable
     {
         return "Edit {$this->record->title}";
-    }
-
-    public function form(Form $form): Form
-    {
-        return $form
-            ->schema([
-                TextInput::make('type')
-                    ->disabled()
-                    ->helperText('This field is not editable.'),
-                TextInput::make('title')
-                    ->required(),
-                Fieldset::make('Content')
-                    ->schema([
-                        Textarea::make('description')
-                            ->required()
-                            ->columnSpan('full'),
-                        Textarea::make('body')
-                            ->required()
-                            ->rows(5)
-                            ->columnSpan('full'),
-                    ]),
-            ]);
-    }
-
-    protected function getFormActions(): array
-    {
-        return [
-            Action::make('Save Changes')
-                ->requiresConfirmation()
-                ->modalIconColor('warning')
-                ->modalHeading("Save Changes to {$this->record->title}?")
-                ->modalDescription($this->record->type->getModalDescription())
-                ->modalSubmitActionLabel('I understand, save changes')
-                ->modalWidth('xl')
-                ->action(function () {
-                    $this->save();
-
-                    if ($this->record->users->count() > 0) {
-                        UserConsentAgreement::where('consent_agreement_id', $this->record->id)
-                            ->delete();
-                    }
-                }),
-            $this->getCancelFormAction(),
-        ];
     }
 }
