+++ resources/views/assist/assistant/filament/pages/assistant-configuration.blade.php
@@ -1,0 +1,3 @@
+<x-filament-panels::page>
+
+</x-filament-panels::page>
